pyproject.toml
[tool.poetry]
name = "gino-admin"
version = "0.2.2"
description = "Admin Panel for PostgreSQL DB with Gino ORM"
authors = ["xnuinside <xnuinside@gmail.com>"]
license = "MIT"
readme = "docs/README.rst"
homepage = "https://github.com/xnuinside/gino-admin"
repository = "https://github.com/xnuinside/gino-admin"

[tool.poetry.dependencies]
python = "^3.7"
Sanic-Jinja2 = "0.8.0"
passlib = "^1.7.4"
aiofiles = "^0.5.0"
PyYAML = "^5.3.1"
gino = "^1.0.1"
click = "^7.1.2"
sanic-jwt = "^1.4.1"
pydantic = "^1.5.1"
dsnparse = "^0.1.15"
expiring-dict = "^1.1.0"
sqlalchemy_utils = "^0.36.8"

[tool.poetry.dev-dependencies]
pre-commit = "^2.3.0"
flake8 = "^3.7.9"
pytest-docker-compose = "^3.1.2"
pytest = "^5.4.2"
sphinx = "^3.0.3"
m2r = "^0.2.1"
twine = "^3.2.0"

[tool.poetry.scripts]
gino-admin = 'gino_admin.cli:cli'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
